Fixes issue introduced in #4073.   * Moves ListRows tests to correct files.   * Adds testing for ListRowsAsync.
